[FLINK-25888][metrics] Move JobStatusMetrics to upper level
