+++ flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/SchedulerBase.java
@@ -81,6 +81,7 @@ import org.apache.flink.runtime.query.KvStateLocation;
 import org.apache.flink.runtime.query.UnknownKvStateLocation;
 import org.apache.flink.runtime.scheduler.exceptionhistory.FailureHandlingResultSnapshot;
 import org.apache.flink.runtime.scheduler.exceptionhistory.RootExceptionHistoryEntry;
+import org.apache.flink.runtime.scheduler.metrics.JobStatusMetrics;
 import org.apache.flink.runtime.scheduler.stopwithsavepoint.StopWithSavepointTerminationHandlerImpl;
 import org.apache.flink.runtime.scheduler.stopwithsavepoint.StopWithSavepointTerminationManager;
 import org.apache.flink.runtime.scheduler.strategy.ExecutionVertexID;
@@ -93,8 +94,6 @@ import org.apache.flink.runtime.util.IntArrayList;
 import org.apache.flink.util.ExceptionUtils;
 import org.apache.flink.util.FlinkException;
 import org.apache.flink.util.IterableUtils;
-import org.apache.flink.util.clock.Clock;
-import org.apache.flink.util.clock.SystemClock;
 import org.apache.flink.util.concurrent.FutureUtils;
 
 import org.slf4j.Logger;
@@ -108,7 +107,6 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 import java.util.OptionalLong;
@@ -635,92 +633,6 @@ public abstract class SchedulerBase implements SchedulerNG, CheckpointScheduling
 
         jobStatusListenerRegistrar.accept(
                 new JobStatusMetrics(metrics, initializationTimestamp, jobStatusMetricsSettings));
-    }
-
-    @VisibleForTesting
-    static class JobStatusMetrics implements JobStatusListener {
-
-        private JobStatus currentStatus = JobStatus.INITIALIZING;
-        private long currentStatusTimestamp;
-        private final long[] cumulativeStatusTimes;
-
-        public JobStatusMetrics(
-                MetricGroup metricGroup,
-                long initializationTimestamp,
-                MetricOptions.JobStatusMetricsSettings jobStatusMetricsSettings) {
-
-            currentStatus = JobStatus.INITIALIZING;
-            currentStatusTimestamp = initializationTimestamp;
-            cumulativeStatusTimes = new long[JobStatus.values().length];
-
-            for (JobStatus jobStatus : JobStatus.values()) {
-                if (!jobStatus.isTerminalState() && jobStatus != JobStatus.RECONCILING) {
-
-                    if (jobStatusMetricsSettings.isStateMetricsEnabled()) {
-                        metricGroup.gauge(
-                                getStateMetricName(jobStatus), createStateMetric(jobStatus));
-                    }
-
-                    if (jobStatusMetricsSettings.isCurrentTimeMetricsEnabled()) {
-                        metricGroup.gauge(
-                                getCurrentTimeMetricName(jobStatus),
-                                createCurrentTimeMetric(jobStatus, SystemClock.getInstance()));
-                    }
-
-                    if (jobStatusMetricsSettings.isTotalTimeMetricsEnabled()) {
-                        metricGroup.gauge(
-                                getTotalTimeMetricName(jobStatus),
-                                createTotalTimeMetric(jobStatus, SystemClock.getInstance()));
-                    }
-                }
-            }
-        }
-
-        @VisibleForTesting
-        Gauge<Long> createStateMetric(JobStatus jobStatus) {
-            return () -> currentStatus == jobStatus ? 1L : 0L;
-        }
-
-        @VisibleForTesting
-        Gauge<Long> createCurrentTimeMetric(JobStatus jobStatus, Clock clock) {
-            return () ->
-                    currentStatus == jobStatus
-                            ? Math.max(clock.absoluteTimeMillis() - currentStatusTimestamp, 0)
-                            : 0;
-        }
-
-        @VisibleForTesting
-        Gauge<Long> createTotalTimeMetric(JobStatus jobStatus, Clock clock) {
-            return () ->
-                    currentStatus == jobStatus
-                            ? cumulativeStatusTimes[jobStatus.ordinal()]
-                                    + Math.max(
-                                            clock.absoluteTimeMillis() - currentStatusTimestamp, 0)
-                            : cumulativeStatusTimes[jobStatus.ordinal()];
-        }
-
-        @VisibleForTesting
-        static String getStateMetricName(JobStatus jobStatus) {
-            return jobStatus.name().toLowerCase(Locale.ROOT) + "State";
-        }
-
-        @VisibleForTesting
-        static String getCurrentTimeMetricName(JobStatus jobStatus) {
-            return jobStatus.name().toLowerCase(Locale.ROOT) + "Time";
-        }
-
-        @VisibleForTesting
-        static String getTotalTimeMetricName(JobStatus jobStatus) {
-            return jobStatus.name().toLowerCase(Locale.ROOT) + "TimeTotal";
-        }
-
-        @Override
-        public void jobStatusChanges(JobID jobId, JobStatus newJobStatus, long timestamp) {
-            cumulativeStatusTimes[currentStatus.ordinal()] += timestamp - currentStatusTimestamp;
-
-            currentStatus = newJobStatus;
-            currentStatusTimestamp = timestamp;
-        }
     }
 
     protected abstract void startSchedulingInternal();
